requirements.txt
langchain==0.0.333
#openai==0.28
openai==1.2.3
python-dotenv